Move endpoints into their own extension method classes
