+++ src/Api/Endpoints/Descriptions.cs
@@ -1,0 +1,6 @@
+namespace Defra.TradeImportsReportingApi.Api.Endpoints;
+
+public static class Descriptions
+{
+    public static readonly string SearchablePeriod = $"Searchable period is the last {TimePeriod.MaxDays} days";
+}
