+++ src/Api/Endpoints/TimePeriod.cs
@@ -1,0 +1,6 @@
+namespace Defra.TradeImportsReportingApi.Api.Endpoints;
+
+public static class TimePeriod
+{
+    public const int MaxDays = 122; // Roughly 4 months based on 365/12 x 4
+}
