+++ tests/Api.Tests/Endpoints/GetBucketsTests.cs
@@ -111,11 +111,7 @@ public class GetBucketsTests(ApiWebApplicationFactory factory, ITestOutputHelper
             Testing.Endpoints.Buckets.Get(
                 EndpointQuery
                     .New.Where(EndpointFilter.From(DateTime.UtcNow))
-                    .Where(
-                        EndpointFilter.To(
-                            DateTime.UtcNow.AddDays(EndpointRouteBuilderExtensions.TimePeriod.MaxDays + 1)
-                        )
-                    )
+                    .Where(EndpointFilter.To(DateTime.UtcNow.AddDays(TimePeriod.MaxDays + 1)))
                     .Where(EndpointFilter.Unit(Units.Hour))
             )
         );
