+++ tests/Api.Tests/Endpoints/GetClearanceRequestsIntervalsTests.cs
@@ -97,11 +97,7 @@ public class GetClearanceRequestsIntervalsTests(ApiWebApplicationFactory factory
             Testing.Endpoints.ClearanceRequests.Intervals(
                 EndpointQuery
                     .New.Where(EndpointFilter.From(DateTime.UtcNow))
-                    .Where(
-                        EndpointFilter.To(
-                            DateTime.UtcNow.AddDays(EndpointRouteBuilderExtensions.TimePeriod.MaxDays + 1)
-                        )
-                    )
+                    .Where(EndpointFilter.To(DateTime.UtcNow.AddDays(TimePeriod.MaxDays + 1)))
                     .Where(EndpointFilter.Unit(Units.Hour))
             )
         );
