+++ tests/Api.Tests/Endpoints/GetClearanceRequestsSummaryTests.cs
@@ -78,11 +78,7 @@ public class GetClearanceRequestsSummaryTests(ApiWebApplicationFactory factory, 
             Testing.Endpoints.ClearanceRequests.Summary(
                 EndpointQuery
                     .New.Where(EndpointFilter.From(DateTime.UtcNow))
-                    .Where(
-                        EndpointFilter.To(
-                            DateTime.UtcNow.AddDays(EndpointRouteBuilderExtensions.TimePeriod.MaxDays + 1)
-                        )
-                    )
+                    .Where(EndpointFilter.To(DateTime.UtcNow.AddDays(TimePeriod.MaxDays + 1)))
             )
         );
