+++ tests/Api.Tests/Endpoints/GetIntervalsTests.cs
@@ -136,11 +136,7 @@ public class GetIntervalsTests(ApiWebApplicationFactory factory, ITestOutputHelp
             Testing.Endpoints.Intervals.Get(
                 EndpointQuery
                     .New.Where(EndpointFilter.From(DateTime.UtcNow))
-                    .Where(
-                        EndpointFilter.To(
-                            DateTime.UtcNow.AddDays(EndpointRouteBuilderExtensions.TimePeriod.MaxDays + 1)
-                        )
-                    )
+                    .Where(EndpointFilter.To(DateTime.UtcNow.AddDays(TimePeriod.MaxDays + 1)))
             )
         );
