+++ tests/Api.Tests/Endpoints/GetMatchesBucketsTests.cs
@@ -87,11 +87,7 @@ public class GetMatchesBucketsTests(ApiWebApplicationFactory factory, ITestOutpu
             Testing.Endpoints.Matches.Buckets(
                 EndpointQuery
                     .New.Where(EndpointFilter.From(DateTime.UtcNow))
-                    .Where(
-                        EndpointFilter.To(
-                            DateTime.UtcNow.AddDays(EndpointRouteBuilderExtensions.TimePeriod.MaxDays + 1)
-                        )
-                    )
+                    .Where(EndpointFilter.To(DateTime.UtcNow.AddDays(TimePeriod.MaxDays + 1)))
                     .Where(EndpointFilter.Unit(Units.Hour))
             )
         );
