+++ tests/Api.Tests/Endpoints/GetMatchesIntervalsTests.cs
@@ -97,11 +97,7 @@ public class GetMatchesIntervalsTests(ApiWebApplicationFactory factory, ITestOut
             Testing.Endpoints.Matches.Intervals(
                 EndpointQuery
                     .New.Where(EndpointFilter.From(DateTime.UtcNow))
-                    .Where(
-                        EndpointFilter.To(
-                            DateTime.UtcNow.AddDays(EndpointRouteBuilderExtensions.TimePeriod.MaxDays + 1)
-                        )
-                    )
+                    .Where(EndpointFilter.To(DateTime.UtcNow.AddDays(TimePeriod.MaxDays + 1)))
             )
         );
