+++ tests/Api.Tests/Endpoints/GetMatchesSummaryTests.cs
@@ -78,11 +78,7 @@ public class GetMatchesSummaryTests(ApiWebApplicationFactory factory, ITestOutpu
             Testing.Endpoints.Matches.Summary(
                 EndpointQuery
                     .New.Where(EndpointFilter.From(DateTime.UtcNow))
-                    .Where(
-                        EndpointFilter.To(
-                            DateTime.UtcNow.AddDays(EndpointRouteBuilderExtensions.TimePeriod.MaxDays + 1)
-                        )
-                    )
+                    .Where(EndpointFilter.To(DateTime.UtcNow.AddDays(TimePeriod.MaxDays + 1)))
             )
         );
