+++ tests/Api.Tests/Endpoints/GetNotificationsBucketsTests.cs
@@ -87,11 +87,7 @@ public class GetNotificationsBucketsTests(ApiWebApplicationFactory factory, ITes
             Testing.Endpoints.Notifications.Buckets(
                 EndpointQuery
                     .New.Where(EndpointFilter.From(DateTime.UtcNow))
-                    .Where(
-                        EndpointFilter.To(
-                            DateTime.UtcNow.AddDays(EndpointRouteBuilderExtensions.TimePeriod.MaxDays + 1)
-                        )
-                    )
+                    .Where(EndpointFilter.To(DateTime.UtcNow.AddDays(TimePeriod.MaxDays + 1)))
                     .Where(EndpointFilter.Unit(Units.Hour))
             )
         );
