+++ tests/Api.Tests/Endpoints/GetNotificationsIntervalsTests.cs
@@ -97,11 +97,7 @@ public class GetNotificationsIntervalsTests(ApiWebApplicationFactory factory, IT
             Testing.Endpoints.Notifications.Intervals(
                 EndpointQuery
                     .New.Where(EndpointFilter.From(DateTime.UtcNow))
-                    .Where(
-                        EndpointFilter.To(
-                            DateTime.UtcNow.AddDays(EndpointRouteBuilderExtensions.TimePeriod.MaxDays + 1)
-                        )
-                    )
+                    .Where(EndpointFilter.To(DateTime.UtcNow.AddDays(TimePeriod.MaxDays + 1)))
             )
         );
