+++ tests/Api.Tests/Endpoints/GetReleasesIntervalsTests.cs
@@ -97,11 +97,7 @@ public class GetReleasesIntervalsTests(ApiWebApplicationFactory factory, ITestOu
             Testing.Endpoints.Releases.Intervals(
                 EndpointQuery
                     .New.Where(EndpointFilter.From(DateTime.UtcNow))
-                    .Where(
-                        EndpointFilter.To(
-                            DateTime.UtcNow.AddDays(EndpointRouteBuilderExtensions.TimePeriod.MaxDays + 1)
-                        )
-                    )
+                    .Where(EndpointFilter.To(DateTime.UtcNow.AddDays(TimePeriod.MaxDays + 1)))
             )
         );
