+++ tests/Api.Tests/Endpoints/GetReleasesSummaryTests.cs
@@ -78,11 +78,7 @@ public class GetReleasesSummaryTests(ApiWebApplicationFactory factory, ITestOutp
             Testing.Endpoints.Releases.Summary(
                 EndpointQuery
                     .New.Where(EndpointFilter.From(DateTime.UtcNow))
-                    .Where(
-                        EndpointFilter.To(
-                            DateTime.UtcNow.AddDays(EndpointRouteBuilderExtensions.TimePeriod.MaxDays + 1)
-                        )
-                    )
+                    .Where(EndpointFilter.To(DateTime.UtcNow.AddDays(TimePeriod.MaxDays + 1)))
             )
         );
