+++ tests/Api.Tests/Endpoints/GetSummaryTests.cs
@@ -87,11 +87,7 @@ public class GetSummaryTests(ApiWebApplicationFactory factory, ITestOutputHelper
             Testing.Endpoints.Summary.Get(
                 EndpointQuery
                     .New.Where(EndpointFilter.From(DateTime.UtcNow))
-                    .Where(
-                        EndpointFilter.To(
-                            DateTime.UtcNow.AddDays(EndpointRouteBuilderExtensions.TimePeriod.MaxDays + 1)
-                        )
-                    )
+                    .Where(EndpointFilter.To(DateTime.UtcNow.AddDays(TimePeriod.MaxDays + 1)))
             )
         );
